requirements/local.txt
-r tests.txt

django-debug-toolbar==3.2.1
gevent==21.1.2
